Add files via upload
This program is used to modify a target network interfaces' MAC address using user input.

The program makes use of the 'subprocess' module to allow for BASH commands to be run from within the program. Additionally, the program allows the user to input a target network interface where the MAC address will be modified.